Move upsell nudge into it's own component
